requirements.txt
Flask
pyspark
mysql-connector-python